examples: Fix some warnings
Fix some warnings in the Basic example. Note that using forever could
make the computation more strict, which would be better - however, the
purpose of the example explicitly seems to be creating space leaks.
Since foldl' is supposed to blow the stack, removing one space leak
seems to be fine.