Report counters in an incremental fashion.
Commit e444d8b680e92c4e8a45319ee41956433025b28d
eliminated the suppression of messages that were
in some sense redundant, but as a side effect
it started reporting the current counter value
instead of the increment since the last report.

This change causes only the increment to be
reported--which is what statsd servers expect.

Gauges are unaffected.  This change affects
only counters, but that includes the counter
within each distribution.